upgrade API calls to Graph API v3.1
Facebook will retire the v2.8 endpoint used in this module on 2018-08-05,
so we must upgrade our calls to use the new latest version of the API,
v3.1. We also had to make a few minor changes to fit the version
upgrade:

* remove the warning when the version is about to expire, since v3.1
still has no end of life date.

* upgrade "state" to a required argument on the API. To avoid breaking
code, we use time() to create our own if no state argument is provided.

* updated documentation, including the removal of some deprecated scope
values.

https://developers.facebook.com/docs/facebook-login
https://developers.facebook.com/blog/post/2018/07/26/introducing-graph-api-v3.1-marketing-api-v3.1-graph-api-explorer-beta/
https://developers.facebook.com/docs/graph-api/changelog/version3.1