api 3.1 upgrade (#16)
* upgrade API calls to Graph API v3.1

Facebook will retire the v2.8 endpoint used in this module on 2018-08-05,
so we must upgrade our calls to use the new latest version of the API,
v3.1. We also had to make a few minor changes to fit the version
upgrade:

* remove the warning when the version is about to expire, since v3.1
still has no end of life date.

* upgrade "state" to a required argument on the API. To avoid breaking
code, we use time() to create our own if no state argument is provided.

* updated documentation, including the removal of some deprecated scope
values.

https://developers.facebook.com/docs/facebook-login
https://developers.facebook.com/blog/post/2018/07/26/introducing-graph-api-v3.1-marketing-api-v3.1-graph-api-explorer-beta/
https://developers.facebook.com/docs/graph-api/changelog/version3.1

* introduce the new debug_token() method

The new standard way of inspecting an access token from Facebook
is to call the debug_token route from the Graph API, which provides
us with a user_id, validation data, accepted scope and other
important information. The debug_token() method is thus introduced,
to simplify access token validation.

https://developers.facebook.com/docs/facebook-login/manually-build-a-login-flow#confirm

* update documentation

* allow custom api version setting

It is sometimes usefull to simply change the API version number instead
of passing the entire auth urls to the constructor.

https://github.com/mamod/Net-Facebook-Oauth2/issues/15

* minor debug_token fixes

* ability to upgrade access token to long-lived

* minor doc rewording

* add missing dep

* sync README to the current docs

* minor documentation updates

* update VERSION and Changes for new release

* doc improvements